backend/requirements.txt
# Web framework
fastapi==0.115.4
uvicorn[standard]==0.32.0


# Machine Learning / NLP
torch==2.2.2
sentence-transformers
faiss-cpu==1.7.4
rank-bm25==0.2.2

# HuggingFace Hub (latest compatible)
huggingface-hub

# Transformers stack (let pip resolve compatible versions)


# Utilities
diskcache==5.6.3
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0

# LLM API client
groq==0.9.0

# Fix compatibility
numpy<2.0
